requirements.txt
torch
runway-python==0.5.4
transformers 
# progress bars in model download and training scripts
tqdm
# Accessing files from S3 directly.
boto3
# Used for downloading models over HTTP
requests
# For XLNet
sentencepieces